FILE: moves createModel
